0.98版本 修改了枚举类不能显示的问题

+++ src/main/java/com/lz/pojo/dto/TaskDraftDTO.java
@@ -20,7 +20,7 @@ import java.util.Date;
 @Data
 public class TaskDraftDTO {
     Long taskId;
-    Integer type;
+    Long type;
     String location;
     String description;
     Date createdAt;
